pyproject.toml
[tool.poetry]
name = "homestock"
version = "2024.12"
description = "Household food stock management"
authors = ["Antonio Jiménez <antoniojr997@gmail.com>"]
license = "GPL-3.0-or-later"

[tool.poetry.dependencies]
python = "^3.9.21"
PyYAML = "6.0.1"
SQLAlchemy = "2.0.36"
python-dotenv = "1.0.1"
pymysql = "1.1.1"
debugpy = "1.8.9"
cryptography = "44.0.0"
pika = "1.3.2"
flask = "3.1.0"
authlib = "1.4.0"
google-auth = "2.38.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"